Some changes: - ultra[4] commented - new pid coeff's - engine saving variables get back to double instead of int's

+++ Pomidor15/src/settings.c
@@ -52,6 +52,6 @@ double settingDistanceToRunFromEnemy = 25;
  */
 int settingTimeBlockedToFail = 600;
 
-int settingVelocityToBlocked = 30;
+double settingVelocityToBlocked = 30;
 
-int settingPWMToBlocked = 0.25;
+double settingPWMToBlocked = 0.25;
